test: Parallelize more!
